cleaning code and organizing classes
