+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapBuilder.java
@@ -21,7 +21,6 @@ class GoogleMapBuilder implements GoogleMapOptionsSink {
   private boolean indoorEnabled = true;
   private boolean trafficEnabled = false;
   private boolean buildingsEnabled = true;
-  private boolean markerCachingEnabled = false;
   private boolean markersAnimationEnabled = false;
   private Object initialMarkers;
   private Object initialPolygons;
@@ -50,7 +49,6 @@ class GoogleMapBuilder implements GoogleMapOptionsSink {
     controller.setInitialCircles(initialCircles);
     controller.setPadding(padding.top, padding.left, padding.bottom, padding.right);
     controller.setInitialTileOverlays(initialTileOverlays);
-    controller.setMarkerCachingEnabled(markerCachingEnabled);
     controller.setMarkersAnimationEnabled(markersAnimationEnabled);
     return controller;
   }
@@ -177,11 +175,6 @@ class GoogleMapBuilder implements GoogleMapOptionsSink {
   @Override
   public void setInitialTileOverlays(List<Map<String, ?>> initialTileOverlays) {
     this.initialTileOverlays = initialTileOverlays;
-  }
-
-  @Override
-  public void setMarkerCachingEnabled(boolean isEnabled) {
-    this.markerCachingEnabled = isEnabled;
   }
 
   @Override
