+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapOptionsSink.java
@@ -56,7 +56,5 @@ interface GoogleMapOptionsSink {
 
   void setInitialTileOverlays(List<Map<String, ?>> initialTileOverlays);
 
-  void setMarkerCachingEnabled(boolean toBoolean);
-
   void setMarkersAnimationEnabled(boolean markersAnimationEnabled);
 }
