+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/MarkerController.java
@@ -8,6 +8,8 @@ import com.google.android.gms.maps.model.BitmapDescriptor;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
 
+import io.flutter.plugins.googlemaps.cozy.CozyMarkerData;
+
 /** Controller of a single Marker on the map. */
 class MarkerController implements MarkerOptionsSink {
