+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/MarkersController.java
@@ -27,6 +27,7 @@ import java.util.Objects;
 import java.util.*;
 import android.os.*;
 
+import io.flutter.plugins.googlemaps.cozy.*;
 import io.flutter.plugin.common.MethodChannel;
 
 class MarkersController {
