+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/cozy/CozyMarkerData.java
@@ -1,4 +1,4 @@
-package io.flutter.plugins.googlemaps;
+package io.flutter.plugins.googlemaps.cozy;
 
 public class CozyMarkerData {
     public final String label;
