+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/cozy/CozyMarkerElement.java
@@ -1,15 +1,15 @@
-package io.flutter.plugins.googlemaps;
+package io.flutter.plugins.googlemaps.cozy;
 
 import android.graphics.RectF;
 
-public class CozyMarkerObject {
+public class CozyMarkerElement {
     public final RectF bounds;
     public final int fillColor;
     public final int strokeColor;
     public final float alpha;
     public Object data;
 
-    public CozyMarkerObject(RectF bounds, int fillColor, int strokeColor, float alpha, Object data) {
+    public CozyMarkerElement(RectF bounds, int fillColor, int strokeColor, float alpha, Object data) {
         this.bounds = bounds;
         this.fillColor = fillColor;
         this.strokeColor = strokeColor;
@@ -17,23 +17,23 @@ public class CozyMarkerObject {
         this.data = data;
     }
 
-    public CozyMarkerObject(RectF bounds, int fillColor, int strokeColor, float alpha) {
+    public CozyMarkerElement(RectF bounds, int fillColor, int strokeColor, float alpha) {
         this(bounds, fillColor, strokeColor, alpha, null);
     }
 
-    public CozyMarkerObject(RectF bounds, int fillColor, int strokeColor) {
+    public CozyMarkerElement(RectF bounds, int fillColor, int strokeColor) {
         this(bounds, fillColor, strokeColor, 1.0f, null);
     }
 
-    public CozyMarkerObject(RectF bounds, int fillColor, float alpha, Object data) {
+    public CozyMarkerElement(RectF bounds, int fillColor, float alpha, Object data) {
         this(bounds, fillColor, 0, alpha, data);
     }
 
-    public CozyMarkerObject(RectF bounds, int fillColor, float alpha) {
+    public CozyMarkerElement(RectF bounds, int fillColor, float alpha) {
         this(bounds, fillColor, 0, alpha, null);
     }
 
-    public CozyMarkerObject(RectF bounds) {
+    public CozyMarkerElement(RectF bounds) {
         this(bounds, 0, 0, 1.0f, null);
     }
